requirements_versions.txt

torchsde==0.2.6
einops==0.8.0
safetensors==0.4.3
accelerate==0.32.1
pyyaml==6.0.1
pillow==10.4.0
scipy==1.14.0
tqdm==4.66.4
psutil==6.0.0
pytorch_lightning==2.3.3
omegaconf==2.3.0
gradio==3.41.2
pygit2==1.15.1
opencv-contrib-python-headless==4.10.0.84
httpx==0.27.0
#onnxruntime-gpu==1.22.0
onnxruntime==1.18.1
timm==1.0.7
numpy==1.26.4
tokenizers==0.19.1
packaging==24.1
rembg==2.0.57
groundingdino-py==0.4.0
segment_anything==1.0
piexif==1.1.3
sentencepiece==0.2.0
opencv-python==4.8.1.78
chardet==5.2.0
insightface==0.7.3
pydantic==2.11.5
lpips==0.1.4
peft==0.6.0
diffusers==0.29.0
##huggingface-hub==0.23.2
##huggingface-hub==0.30.0
controlnet_aux==0.0.10
potracer
deep-translator
ultralytics==8.3.227
mediapipe==0.10.13
transformers==4.42.4
